requirements.txt
boto3
requests
python-dotenv
botocore
camelot-py
cloudscraper
cssselect
Jinja2
jpype1
lxml
more-itertools
numpy
opencv-python-headless
openpyxl
pandas
pandera
parsel
pdfminer.six
pdfplumber
pillow
playwright
pydantic
pydantic_core
pypdf
PyPDF2
pypdfium2
python-dateutil
python-docx
python-dotenv
Scrapy
scrapy-inline-requests
tabula
tabula-py
urllib3
